wip: mvlc readout parsing

+++ src/databuffer.h
@@ -31,7 +31,7 @@ struct DataBuffer
         : DataBuffer(0, 0, 0)
     {}
 
-    DataBuffer(size_t sz, int tag = 0, u64 id = 0u)
+    DataBuffer(size_t sz, int tag = 0, u32 id = 0u)
         : data(nullptr)
         , size(sz)
         , used(0)
@@ -161,7 +161,7 @@ struct DataBuffer
     u8 *data;
     size_t size; // size in bytes
     size_t used; // bytes used
-    u64 id = 0u; // id value for external use
+    u32 id = 0u; // id value for external use
     int tag = 0; // tag allowing to distinguish buffer types
 };
